change .toDouble to .0 for shorter syntax

+++ lib/constants.dart
@@ -5,6 +5,6 @@
 // it have a max brightness up to 2047 so I'll put this here for now, in
 // hope that the android team can add this feature
 final Map<String, double> maxBrightnessMap = {
-  "Xiaomi M2101K6G": 2047.toDouble(),
-  "default": 255.toDouble()
+  "Xiaomi M2101K6G": 2047.0,
+  "default": 255.0
 };
